🚚 Move Component4069 implementation and tests
