+++ src/Factory/ComponentFactory.cpp
@@ -20,7 +20,7 @@
 #include "Component4001.hpp"
 #include "Component4011.hpp"
 #include "Component4030.hpp"
-// #include "Component4069.hpp"
+#include "Component4069.hpp"
 #include "Component4071.hpp"
 #include "Component4081.hpp"
 
@@ -39,7 +39,7 @@ nts::ComponentFactory::ComponentFactory()
         {"4001", []() { return std::make_unique<nts::Component4001>("4001"); }},
         {"4011", []() { return std::make_unique<nts::Component4011>("4011"); }},
         {"4030", []() { return std::make_unique<nts::Component4030>("4030"); }},
-        // {"4069", []() { return std::make_unique<nts::Component4069>("4069"); }},
+        {"4069", []() { return std::make_unique<nts::Component4069>("4069"); }},
         {"4071", []() { return std::make_unique<nts::Component4071>("4071"); }},
         {"4081", []() { return std::make_unique<nts::Component4081>("4081"); }}
     };
